EPBR-6410 Allow enumerable data to be saved for RdSAP 21
This fixes the issue in which the ExternalDefintions.xml could not be read by read_xml method in the ViewModel gem. This was due to the use of  Nokogiri config.huge.strict. The fix uses inheritance to override the existing read_xml method to not specify this Nokogiri config option rather updating this method used as part of lodgement