decreases RAM use in export assessment documents
The existing export_documents rake can run out of memory when run in AWS on large result sets with jsonb. This updates the export process to extract only assessment_id and loop over this result and extract each jsonb in a single db transaction. This should limit the RAM usage of the task.