OD-20 updates the audit log insert to use an Upsert
The previous commit allowed dupes for assessment_id and event_type. As far as we can tell this is not fault with insert but rather how the queues are processed. To ensure we have none these dupes I have created a combined constraint on the assessment_id, event_type columns. I have updated the data to insert into this table and ON CONFLICT, update the timestamp to be updated for that row.
The migration for the constraint must first truncate the table before being applied as the existing data will raise an error.
I have also added a primary key to the table