Refactor how enums/lookups are loaded for tests
In order to get enum data in the test db we used the use case that reads the data from the XSDs. This is expensive, to speed up this process, the data is now loaded from a csv. Only data needed for the test is loaded as part of the setup. To produce the csv I loaded all the enums into the dev db using the rake import_enums_xsd, then downloaded the saved data saved into the csv used by the tests.