OD-185 removes the sequence from domestic_rr mview
We added the sequence value to the mview so it had a unique identifier. This is a requirement for doing concurrent refreshes. We no longer needs. The previous commit we removed this value for the view as is not needed however this means that the Iceberg catalog cannot be updated as the columns do not match. This fix that issue by ensuring the sequence is removed