OD-244 Updates certificate endpoint to return json
The existing endpoint was returning a json string rather than json. The endpoint was performing a 'to_json' on object that  had already been converted to json by the redacted class and is further converted into json by the method 'json_api_response'.

Fixes by updating the redacted class to keep the document as hash
The hash is converted to json by the controller only.

Adds performance improvements to the document gateway and removes the assessment_id as a return column.

Adds performance improvements to redacted class by passing the json as a hash with symbolized keys