OD-176 updates commercial rr sql to use assessment_documents
The existing mview took 3 days to run. This is because as it had to scan the EAV table for multiple attributes and EPCs for each CEPC-RR . As there is no variation in CEPC-RR across schema types there is no need to use this table. Using the assessment_documents gives better query performance.

Updates the view to match the sql which also improves query performance

Adds Related-RRN to xml samples which is now required to make the test pass and provides a closer match to production data